pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mh-mind"
version = "0.1.0"
description = "Personal RAG app for chatting with Apple Notes and Word documents"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "sqlite-vec>=0.1.6",
    "python-docx>=1.1.0",
    "streamlit>=1.38.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
embed = [
    "sentence-transformers>=3.0.0",
    "einops>=0.8.0",
]

[project.scripts]
mh-mind = "mh_mind.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["mh_mind"]
